requirements.txt
python-telegram-bot==22.0
yt-dlp>=2025.04.30
python-dotenv>=1.1.0
filelock>=3.16.1
spotdl>=4.2.11